VTKU-31 : Split postprocessor in two
No functional changes.
