+++ sijoittelu-algoritmi-batch/src/main/java/fi/vm/sade/sijoittelu/batch/logic/impl/algorithm/postsijoitteluprocessor/PostSijoitteluProcessor.java
@@ -13,6 +13,7 @@ public interface PostSijoitteluProcessor extends Processor {
 
     static Collection<PostSijoitteluProcessor> defaultPostProcessors() {
         return ImmutableList.of(
+                new PostSijoitteluProcessorKorjaaHakemustenTilatJonoissaJoissaEiOleVarasijatayttoaSivssnovinJalkeen(),
                 new PostSijoitteluProcessorPeruunnutaRajatunVarasijataytonHakemuksetJotkaEivatMahduVaralle(),
                 new PostSijoitteluProcessorPeruunnutaYlijaaneetVarallaTilaisetJosVarasijatayttoOnPaattynyt(),
                 new PostSijoitteluProcessorAsetaSivssnov(),
