+++ sijoittelu-algoritmi-batch/src/main/java/fi/vm/sade/sijoittelu/batch/logic/impl/algorithm/postsijoitteluprocessor/SivssnovUtil.java
@@ -1,0 +1,33 @@
+package fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.postsijoitteluprocessor;
+
+import fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.wrappers.SijoitteluajoWrapper;
+import fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.wrappers.ValintatapajonoWrapper;
+import fi.vm.sade.sijoittelu.domain.Valintatapajono;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class SivssnovUtil {
+    private static final Logger LOG = LoggerFactory.getLogger(SivssnovUtil.class);
+
+    public static void assertSijoiteltuEnnenVarasijataytonLoppumista(ValintatapajonoWrapper jonoWrapper, SijoitteluajoWrapper sijoitteluajoWrapper) {
+        Valintatapajono jono = jonoWrapper.getValintatapajono();
+        if (!jono.getSijoiteltuIlmanVarasijasaantojaNiidenOllessaVoimassa()
+            && sijoitteluajoWrapper.onkoVarasijaSaannotVoimassaJaVarasijaTayttoPaattynyt(jonoWrapper)) {
+                throw new IllegalStateException("Haun " + sijoitteluajoWrapper.getSijoitteluajo().getHakuOid() + " hakukohteen " +
+                    jonoWrapper.getHakukohdeWrapper().getHakukohde().getOid() + " valintatapajonoa " + jono.getOid() +
+                    " ei ole kertaakaan sijoiteltu ilman varasijasääntöjä niiden ollessa voimassa, vaikka sen varasijatäyttö " +
+                    "on jo päättynyt.");
+        }
+    }
+
+    public static boolean varasijaSaannotVoimassa(SijoitteluajoWrapper ajoWrapper) {
+        LOG.debug("Haun " + ajoWrapper.getSijoitteluajo().getHakuOid() + " varasijasäännöt voimassa: " +
+            ajoWrapper.getVarasijaSaannotAstuvatVoimaan() + " -- " + ajoWrapper.getVarasijaTayttoPaattyy());
+        if (!ajoWrapper.varasijaSaannotVoimassa()) {
+            LOG.debug("Haun " + ajoWrapper.getSijoitteluajo().getHakuOid() +
+                " varasijasäännöt eivät ole voimassa.");
+            return false;
+        }
+        return true;
+    }
+}
